requirements.txt
Flask
speechrecognition
python-docx
gunicorn
openpyxl